// src/app/types/generated/account-create-request-dto.ts
// ---------------------------------------------------------------------------
// GENERATED — do not edit manually.
// Source: com.example.planeo_back.web.DTO.account.AccountCreateRequestDTO
// Generator: ts-codegen-processor (https://github.com/your-org/ts-codegen)
// ---------------------------------------------------------------------------

export interface AccountCreateRequestDTO {
  label: string;
  amount: number;
  logo: string;
}